Rename koiki.models to koiki.resources (#93)
* rename koiki.models to koiki.resources

* rename koiki.woocommerce.models to koiki.woocommerce.resources
